OpenGL Cache: Rearrange methods
When splitting the file into different commits, I tried to get a good
diff by putting them next to their old counterparts. This kept it from
compiling since some static methods were near the bottom and needed to
be at the top of the file. From this commit on, it should be compiling
with every commit.