CMake: Add custom command to generate scm_rev
Adds a new field to scm_rev "shader_cache_version" which is dependent on
several files. In order to keep from regenerating the hash too often,
using a custom command we can specify which files affect the hash and
cmake will only rerun the custom command if one of these files change.