Fix dialog and message box placement for SBS 3D mode.
In 3D SBS mode, dialogs should all appear in the left eye instead of split over both eyes.
A 3D-aware QMessageBox3D class is added which should be used instead of QMessageBox.
I didn't change the file dialogs because they already always appeared in the left eye.

Make message boxes (but not dialog boxes) actually stereoscopic with pop-out.

You can still only click on the box in the left eye, which is modal, but a matching inactive box is drawn in the right eye. The message boxes are offset by 8 pixels in each direction so it pops out of the screen. Note that there will be some minor retinal rivalry because the operating system draws active and inactive dialog boxes slightly differently.

This trick is impossible with actual dialog boxes, because their contents can change. So they are still only drawn in the left eye.