OpenGL: Clean up presentation and fix renderdoc
The frontend code surrounding loading OpenGL is sorta complicated for no
reason at the moment. This refactors the code to use a regular Qt Widget
backed by a native window handle to render into.

This change also includes a fix for renderdoc submission. With this
change, when renderdoc or nvidia nsight is detected, the code will
present the final framebuffer from the emuthread, which prevents issues
with multiple context support in renderdoc.